refactor: move CC command handlers out of Node class
