+++ packages/zwave-js/src/lib/driver/Driver.ts
@@ -22,6 +22,7 @@ import {
 	Security2CCCommandsSupportedGet,
 	Security2CCCommandsSupportedReport,
 	Security2CCMessageEncapsulation,
+	Security2CCNonceGet,
 	Security2CCNonceReport,
 	Security2Command,
 	SecurityCC,
@@ -29,6 +30,8 @@ import {
 	SecurityCCCommandEncapsulationNonceGet,
 	SecurityCCCommandsSupportedGet,
 	SecurityCCCommandsSupportedReport,
+	SecurityCCNonceGet,
+	SecurityCCNonceReport,
 	SecurityCommand,
 	SupervisionCC,
 	type SupervisionCCGet,
@@ -92,7 +95,9 @@ import {
 	type ValueMetadata,
 	ZWaveError,
 	ZWaveErrorCodes,
+	allCCs,
 	deserializeCacheValue,
+	encapsulationCCs,
 	generateECDHKeyPair,
 	getCCName,
 	isEncapsulationCC,
@@ -214,6 +219,7 @@ import path from "pathe";
 import { PACKAGE_NAME, PACKAGE_VERSION } from "../_version.js";
 import { ZWaveController } from "../controller/Controller.js";
 import { InclusionState, RemoveNodeReason } from "../controller/Inclusion.js";
+import { determineNIF } from "../controller/NodeInformationFrame.js";
 import { DriverLogger } from "../log/Driver.js";
 import type { Endpoint } from "../node/Endpoint.js";
 import type { ZWaveNode } from "../node/Node.js";
@@ -224,10 +230,10 @@ import {
 	type ZWaveNotificationCallback,
 	zWaveNodeEvents,
 } from "../node/_Types.js";
-import { type ZWaveNodeBase } from "../node/mixins/00_Base.js";
-import { type NodeWakeup } from "../node/mixins/30_Wakeup.js";
-import { type NodeValues } from "../node/mixins/40_Values.js";
-import { type NodeSchedulePoll } from "../node/mixins/60_ScheduledPoll.js";
+import { type ZWaveNodeBase } from "../node/mixins/000_Base.js";
+import { type NodeWakeup } from "../node/mixins/030_Wakeup.js";
+import { type NodeValues } from "../node/mixins/040_Values.js";
+import { type NodeSchedulePoll } from "../node/mixins/060_ScheduledPoll.js";
 import { reportMissingDeviceConfig } from "../telemetry/deviceConfig.js";
 import {
 	type AppInfo,
@@ -2420,6 +2426,9 @@ export class Driver extends TypedEventTarget<DriverEventCallbacks>
 				return { type: "requeue" };
 			});
 		}
+
+		// Start the timer for sending the node to sleep again
+		this.debounceSendNodeToSleep(node);
 	}
 
 	/** Is called when a node goes to sleep */
@@ -3867,7 +3876,10 @@ export class Driver extends TypedEventTarget<DriverEventCallbacks>
 					msg.command
 						instanceof SecurityCCCommandEncapsulationNonceGet
 				) {
-					void this.tryGetNode(msg)?.handleSecurityNonceGet();
+					const node = this.tryGetNode(msg);
+					if (node) {
+						void this.handleSecurityNonceGet(node);
+					}
 				}
 
 				// Transport Service commands must be handled before assembling partial CCs
@@ -5353,6 +5365,34 @@ ${handlers.length} left`,
 			const node = this.controller.nodes.get(nodeId)!;
 			const nodeSessions = this.nodeSessions.get(nodeId);
 			// Check if we need to handle the command ourselves
+
+			// Some Security-related commands make sense to be handled in the driver
+			if (msg.command instanceof SecurityCCNonceGet) {
+				return this.handleSecurityNonceGet(node);
+			}
+			if (msg.command instanceof SecurityCCNonceReport) {
+				return this.handleSecurityNonceReport(node, msg.command);
+			}
+			if (msg.command instanceof SecurityCCCommandsSupportedGet) {
+				return this.handleSecurityCommandsSupportedGet(
+					node,
+					msg.command,
+				);
+			}
+
+			if (msg.command instanceof Security2CCNonceGet) {
+				return this.handleSecurity2NonceGet(node);
+			}
+			if (msg.command instanceof Security2CCNonceReport) {
+				return this.handleSecurity2NonceReport(node, msg.command);
+			}
+			if (msg.command instanceof Security2CCCommandsSupportedGet) {
+				return this.handleSecurity2CommandsSupportedGet(
+					node,
+					msg.command,
+				);
+			}
+
 			if (
 				msg.command.ccId === CommandClasses.Supervision
 				&& msg.command instanceof SupervisionCCReport
@@ -5373,99 +5413,109 @@ ${handlers.length} left`,
 				if (!msg.command.moreUpdatesFollow) {
 					nodeSessions.supervision.delete(msg.command.sessionId);
 				}
+
+				return;
+			}
+
+			// Figure out if the command was received with supervision encapsulation and we need to respond accordingly
+			const supervisionSessionId = SupervisionCC.getSessionId(
+				msg.command,
+			);
+			// Figure out if this is an S2 multicast followup for a group that is out of sync
+			const s2MulticastOutOfSync = this.mustReplyWithSecurityS2MOS(
+				msg,
+			);
+
+			const encapsulationFlags = msg.command.encapsulationFlags;
+
+			let reply: (
+				status:
+					| SupervisionStatus.Success
+					| SupervisionStatus.Fail
+					| SupervisionStatus.NoSupport,
+			) => Promise<void>;
+			if (supervisionSessionId != undefined) {
+				// The command was supervised, and we must respond with a Supervision Report
+				const endpoint = node.getEndpoint(msg.command.endpointIndex)
+					?? node;
+				reply = (status) =>
+					endpoint
+						.createAPI(CommandClasses.Supervision, false)
+						.withOptions({ s2MulticastOutOfSync })
+						.sendReport({
+							sessionId: supervisionSessionId,
+							moreUpdatesFollow: false,
+							status,
+							requestWakeUpOnDemand: this
+								.shouldRequestWakeupOnDemand(node),
+							encapsulationFlags,
+							lowPriority: this
+								.shouldUseLowPriorityForSupervisionReport(
+									node,
+									encapsulationFlags,
+								),
+						});
 			} else {
-				// Figure out if the command was received with supervision encapsulation and we need to respond accordingly
-				const supervisionSessionId = SupervisionCC.getSessionId(
-					msg.command,
-				);
-				// Figure out if this is an S2 multicast followup for a group that is out of sync
-				const s2MulticastOutOfSync = this.mustReplyWithSecurityS2MOS(
-					msg,
-				);
+				// Unsupervised, reply is a no-op
+				reply = () => Promise.resolve();
+			}
+			// DO NOT force-add support for the Supervision CC here. Some devices only support Supervision when sending,
+			// so we need to trust the information we already have.
 
-				const encapsulationFlags = msg.command.encapsulationFlags;
+			// In the case where the command was unsupervised and we need to send a MOS, do it as soon as possible
+			if (supervisionSessionId == undefined && s2MulticastOutOfSync) {
+				// If the command was NOT received using Supervision,
+				// we need to respond with an MOS nonce. Otherwise we'll set the flag
+				// on the Supervision Report
+				node.commandClasses["Security 2"].sendMOS().catch(() => {
+					// Ignore errors
+				});
+			}
 
-				let reply: (
-					status:
-						| SupervisionStatus.Success
-						| SupervisionStatus.Fail
-						| SupervisionStatus.NoSupport,
-				) => Promise<void>;
-				if (supervisionSessionId != undefined) {
-					// The command was supervised, and we must respond with a Supervision Report
-					const endpoint = node.getEndpoint(msg.command.endpointIndex)
-						?? node;
-					reply = (status) =>
-						endpoint
-							.createAPI(CommandClasses.Supervision, false)
-							.withOptions({ s2MulticastOutOfSync })
-							.sendReport({
-								sessionId: supervisionSessionId,
-								moreUpdatesFollow: false,
-								status,
-								requestWakeUpOnDemand: this
-									.shouldRequestWakeupOnDemand(node),
-								encapsulationFlags,
-								lowPriority: this
-									.shouldUseLowPriorityForSupervisionReport(
-										node,
-										encapsulationFlags,
-									),
-							});
-				} else {
-					// Unsupervised, reply is a no-op
-					reply = () => Promise.resolve();
-				}
-				// DO NOT force-add support for the Supervision CC here. Some devices only support Supervision when sending,
-				// so we need to trust the information we already have.
+			// check if someone is waiting for this command
+			for (const entry of this.awaitedCommands) {
+				if (entry.predicate(msg.command)) {
+					// there is!
+					entry.handler(msg.command);
 
-				// In the case where the command was unsupervised and we need to send a MOS, do it as soon as possible
-				if (supervisionSessionId == undefined && s2MulticastOutOfSync) {
-					// If the command was NOT received using Supervision,
-					// we need to respond with an MOS nonce. Otherwise we'll set the flag
-					// on the Supervision Report
-					node.commandClasses["Security 2"].sendMOS().catch(() => {
-						// Ignore errors
-					});
-				}
-
-				// check if someone is waiting for this command
-				for (const entry of this.awaitedCommands) {
-					if (entry.predicate(msg.command)) {
-						// there is!
-						entry.handler(msg.command);
-
-						// and possibly reply to a supervised command
-						await reply(SupervisionStatus.Success);
-						return;
-					}
-				}
-
-				// No one is waiting, dispatch the command to the node itself
-				try {
-					await node.handleCommand(msg.command);
+					// and possibly reply to a supervised command
 					await reply(SupervisionStatus.Success);
-				} catch (e) {
-					let handled = false;
-					if (isZWaveError(e)) {
-						if (e.code === ZWaveErrorCodes.CC_OperationFailed) {
-							// The sending node tried to do something that didn't work
-							await reply(SupervisionStatus.Fail);
-							handled = true;
-						} else if (e.code === ZWaveErrorCodes.CC_NotSupported) {
-							// The sending node sent a command we could not handle
-							await reply(SupervisionStatus.NoSupport);
-							handled = true;
-						}
-					}
+					return;
+				}
+			}
 
-					if (!handled) {
-						// Something unexpected happened.
-						// Report failure, then re-throw the error, so it can be handled accordingly
+			// Some S2 commands contain only extensions. Those are handled by the CC implementation.
+			if (
+				msg.command instanceof Security2CCMessageEncapsulation
+				&& msg.command.encapsulated == undefined
+			) {
+				return;
+			}
+
+			// No one is waiting, dispatch the command to the node itself
+			try {
+				await node.handleCommand(msg.command);
+				await reply(SupervisionStatus.Success);
+			} catch (e) {
+				let handled = false;
+				if (isZWaveError(e)) {
+					if (e.code === ZWaveErrorCodes.CC_OperationFailed) {
+						// The sending node tried to do something that didn't work
 						await reply(SupervisionStatus.Fail);
-
-						throw e;
+						handled = true;
+					} else if (e.code === ZWaveErrorCodes.CC_NotSupported) {
+						// The sending node sent a command we could not handle
+						await reply(SupervisionStatus.NoSupport);
+						handled = true;
 					}
+				}
+
+				if (!handled) {
+					// Something unexpected happened.
+					// Report failure, then re-throw the error, so it can be handled accordingly
+					await reply(SupervisionStatus.Fail);
+
+					throw e;
 				}
 			}
 
@@ -5517,6 +5567,287 @@ ${handlers.length} left`,
 			}
 		} else {
 			this.driverLog.print("  no handlers registered!", "warn");
+		}
+	}
+
+	private hasLoggedNoNetworkKey = false;
+
+	private async handleSecurityNonceGet(
+		node: ZWaveNode,
+	): Promise<void> {
+		// Only reply if secure communication is set up
+		if (!this.securityManager) {
+			if (!this.hasLoggedNoNetworkKey) {
+				this.hasLoggedNoNetworkKey = true;
+				this.controllerLog.logNode(node.id, {
+					message:
+						`cannot reply to NonceGet because no network key was configured!`,
+					direction: "inbound",
+					level: "warn",
+				});
+			}
+			return;
+		}
+
+		// When a node asks us for a nonce, it must support Security CC
+		node.addCC(CommandClasses.Security, {
+			isSupported: true,
+			version: 1,
+			// Security CC is always secure
+			secure: true,
+		});
+
+		// Ensure that we're not flooding the queue with unnecessary NonceReports (GH#1059)
+		const isNonceReport = (t: Transaction) =>
+			t.message.getNodeId() === node.id
+			&& containsCC(t.message)
+			&& t.message.command instanceof SecurityCCNonceReport;
+
+		if (this.hasPendingTransactions(isNonceReport)) {
+			this.controllerLog.logNode(node.id, {
+				message:
+					"in the process of replying to a NonceGet, won't send another NonceReport",
+				level: "warn",
+			});
+			return;
+		}
+
+		// Delete all previous nonces we sent the node, since they should no longer be used
+		this.securityManager.deleteAllNoncesForReceiver(node.id);
+
+		// Now send the current nonce
+		try {
+			await node.commandClasses.Security.sendNonce();
+		} catch (e) {
+			this.controllerLog.logNode(node.id, {
+				message: `failed to send nonce: ${getErrorMessage(e)}`,
+				direction: "inbound",
+			});
+		}
+	}
+
+	/**
+	 * Is called when a nonce report is received that does not belong to any transaction.
+	 * The received nonce reports are stored as "free" nonces
+	 */
+	private handleSecurityNonceReport(
+		node: ZWaveNode,
+		command: SecurityCCNonceReport,
+	): void {
+		const secMan = this.securityManager;
+		if (!secMan) return;
+
+		secMan.setNonce(
+			{
+				issuer: node.id,
+				nonceId: secMan.getNonceId(command.nonce),
+			},
+			{
+				nonce: command.nonce,
+				receiver: this.controller.ownNodeId!,
+			},
+			{ free: true },
+		);
+	}
+
+	private async handleSecurityCommandsSupportedGet(
+		node: ZWaveNode,
+		command: SecurityCCCommandsSupportedGet,
+	): Promise<void> {
+		const endpoint = node.getEndpoint(command.endpointIndex) ?? node;
+
+		if (this.getHighestSecurityClass(node.id) === SecurityClass.S0_Legacy) {
+			const { supportedCCs } = determineNIF();
+			await endpoint.commandClasses.Security.reportSupportedCommands(
+				supportedCCs,
+				// We don't report controlled CCs
+				[],
+			);
+		} else {
+			// S0 is not the highest class. Return an empty list
+			await endpoint.commandClasses.Security.reportSupportedCommands(
+				[],
+				[],
+			);
+		}
+	}
+
+	/** Handles a nonce request for S2 */
+	private async handleSecurity2NonceGet(
+		node: ZWaveNode,
+	): Promise<void> {
+		// Only reply if secure communication is set up
+		if (!this.getSecurityManager2(node.id)) {
+			if (!this.hasLoggedNoNetworkKey) {
+				this.hasLoggedNoNetworkKey = true;
+				this.controllerLog.logNode(node.id, {
+					message:
+						`cannot reply to NonceGet (S2) because no network key was configured!`,
+					direction: "inbound",
+					level: "warn",
+				});
+			}
+			return;
+		}
+
+		// When a node asks us for a nonce, it must support Security 2 CC
+		node.addCC(CommandClasses["Security 2"], {
+			isSupported: true,
+			version: 1,
+			// Security 2 CC is always secure
+			secure: true,
+		});
+
+		// Ensure that we're not flooding the queue with unnecessary NonceReports (GH#1059)
+		const isNonceReport = (t: Transaction) =>
+			t.message.getNodeId() === node.id
+			&& containsCC(t.message)
+			&& t.message.command instanceof Security2CCNonceReport;
+
+		if (this.hasPendingTransactions(isNonceReport)) {
+			this.controllerLog.logNode(node.id, {
+				message:
+					"in the process of replying to a NonceGet, won't send another NonceReport",
+				level: "warn",
+			});
+			return;
+		}
+
+		try {
+			await node.commandClasses["Security 2"].sendNonce();
+		} catch (e) {
+			this.controllerLog.logNode(node.id, {
+				message: `failed to send nonce: ${getErrorMessage(e)}`,
+				direction: "inbound",
+			});
+		}
+	}
+
+	/**
+	 * Is called when a nonce report is received that does not belong to any transaction.
+	 */
+	private handleSecurity2NonceReport(
+		node: ZWaveNode,
+		_command: Security2CCNonceReport,
+	): void {
+		// const secMan = this.securityManager2;
+		// if (!secMan) return;
+
+		// This has the potential of resetting our SPAN state in the middle of a transaction which may expect it to be valid
+		// So we probably shouldn't react here, and instead handle the NonceReport we'll get in response to the next command we send
+
+		// if (command.SOS && command.receiverEI) {
+		// 	// The node couldn't decrypt the last command we sent it. Invalidate
+		// 	// the shared SPAN, since it did the same
+		// 	secMan.storeRemoteEI(node.id, command.receiverEI);
+		// }
+
+		// Since we landed here, this is not in response to any command we sent
+		this.controllerLog.logNode(node.id, {
+			message:
+				`received S2 nonce without an active transaction, not sure what to do with it`,
+			level: "warn",
+			direction: "inbound",
+		});
+	}
+
+	private async handleSecurity2CommandsSupportedGet(
+		node: ZWaveNode,
+		command: Security2CCCommandsSupportedGet,
+	): Promise<void> {
+		const endpoint = node.getEndpoint(command.endpointIndex) ?? node;
+
+		const highestSecurityClass = this.getHighestSecurityClass(node.id);
+		const actualSecurityClass = (
+			command.getEncapsulatingCC(
+				CommandClasses["Security 2"],
+				Security2Command.MessageEncapsulation,
+			) as Security2CCMessageEncapsulation | undefined
+		)?.securityClass;
+
+		if (
+			highestSecurityClass !== undefined
+			&& highestSecurityClass === actualSecurityClass
+		) {
+			// The command was received using the highest security class. Return the list of supported CCs
+
+			const implementedCCs = allCCs.filter((cc) =>
+				getImplementedVersion(cc) > 0
+			);
+
+			// Encapsulation CCs are always supported
+			const implementedEncapsulationCCs = encapsulationCCs.filter(
+				(cc) =>
+					implementedCCs.includes(cc)
+					// A node MUST advertise support for Multi Channel Command Class only if it implements End Points.
+					// A node able to communicate using the Multi Channel encapsulation but implementing no End Point
+					// MUST NOT advertise support for the Multi Channel Command Class.
+					// --> We do not implement end points
+					&& cc !== CommandClasses["Multi Channel"],
+			);
+
+			const supportedCCs = new Set([
+				// DT:00.11.0004.1
+				// All Root Devices or nodes MUST support:
+				// - Association, version 2
+				// - Association Group Information
+				// - Device Reset Locally
+				// - Firmware Update Meta Data, version 5
+				// - Indicator, version 3
+				// - Manufacturer Specific
+				// - Multi Channel Association, version 3
+				// - Powerlevel
+				// - Security 2
+				// - Supervision
+				// - Transport Service, version 2
+				// - Version, version 2
+				// - Z-Wave Plus Info, version 2
+				CommandClasses.Association,
+				CommandClasses["Association Group Information"],
+				CommandClasses["Device Reset Locally"],
+				CommandClasses["Firmware Update Meta Data"],
+				CommandClasses.Indicator,
+				CommandClasses["Manufacturer Specific"],
+				CommandClasses["Multi Channel Association"],
+				CommandClasses.Powerlevel,
+				CommandClasses.Version,
+				CommandClasses["Z-Wave Plus Info"],
+
+				// Generic Controller device type has no additional support requirements,
+				// but we also support the following command classes:
+				CommandClasses["Inclusion Controller"],
+
+				// plus encapsulation CCs, which are part of the above requirement
+				...implementedEncapsulationCCs.filter(
+					(cc) =>
+						// CC:009F.01.0E.11.00F
+						// The Security 0 and Security 2 Command Class MUST NOT be advertised in this command
+						// The Transport Service Command Class MUST NOT be advertised in this command.
+						cc !== CommandClasses.Security
+						&& cc !== CommandClasses["Security 2"]
+						&& cc !== CommandClasses["Transport Service"],
+				),
+			]);
+
+			// Commands that are always in the NIF should not appear in the
+			// S2 commands supported report
+			const commandsInNIF = new Set(determineNIF().supportedCCs);
+			const supportedCommandsNotInNIF = [...supportedCCs].filter((cc) =>
+				!commandsInNIF.has(cc)
+			);
+
+			await endpoint.commandClasses["Security 2"].reportSupportedCommands(
+				supportedCommandsNotInNIF,
+			);
+		} else if (securityClassIsS2(actualSecurityClass)) {
+			// The command was received using a lower security class. Return an empty list
+			await endpoint.commandClasses["Security 2"]
+				.withOptions({
+					s2OverrideSecurityClass: actualSecurityClass,
+				})
+				.reportSupportedCommands([]);
+		} else {
+			// Do not respond
 		}
 	}
