+++ packages/zwave-js/src/lib/node/CCHandlers/_shared.ts
@@ -1,0 +1,1 @@
+export const MAX_ASSOCIATIONS = 1;
