+++ packages/zwave-js/src/lib/node/mixins/050_Endpoints.ts
@@ -11,7 +11,7 @@ import { isArray, isObject } from "alcalzone-shared/typeguards";
 import { DeviceClass } from "../DeviceClass.js";
 import { Endpoint } from "../Endpoint.js";
 import * as nodeUtils from "../utils.js";
-import { NodeCapabilityValuesMixin } from "./41_CapabilityValues.js";
+import { NodeCapabilityValuesMixin } from "./041_CapabilityValues.js";
 
 /** Defines functionality of Z-Wave nodes related to accessing endpoints and their capabilities */
 export interface Endpoints {
